libsvafs: cleanup the directory cache/index
